Separate C extension for Python 2
