+++ cdblib/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
-from .cdblib import djb_hash, Reader, Reader64, Writer, Writer64
+from .djb_hash import djb_hash
+from .cdblib import Reader, Reader64, Writer, Writer64
 
 
 __all__ = ['djb_hash', 'Reader', 'Reader64', 'Writer', 'Writer64']
